Renders .tf files using jinja2 to avoid destroy/recreate actions
When using a count, Terraform identifies resources by the index
of the count. When the user changes the order of the objects in
the list, the details of the indexed resource changes also and so
Terraform thinks every resource has changed. As a result, Terraform
wants to delete/recreate every resource.

This patch works around the problem until it is addressed in
Terraform. Jinja is used to template the .tf files and create
individual resources for every item in the list, with a "lookup"
based on the object to retrieve any associated values.