Ensures resource IDs represent the object being *created*
Previously, the resource ID was the object source, and so objects
would be recreated if the source changed. I.e. if the object were
retrieved from a different source URI, but still written to the
same destination, then the object would be destroyed and recreated
even if the objects contents remained the same. That would lead to
a race condition on the destroy/create cycle.